Move the scope definition to the crate root
